[feature] Improved the scaling of the path precedence algorithm and added test cases

+++ exquery-restxq-api/src/main/java/org/exquery/restxq/annotation/PathAnnotation.java
@@ -74,5 +74,5 @@ public interface PathAnnotation extends ConstraintAnnotation {
      * 
      * @return The path specificity metric
      */
-    public int getPathSpecificityMetric();
+    public long getPathSpecificityMetric();
 }
